budget: colocate Transaction entity into entities/transaction.ts (#563)
Single source of truth for Transaction representations (Domain, IDB,
Raw, Seed) and their parsers/converters/serializers. Adds
entities/_helpers.ts with shared timestamp/date conversion utilities
used by entity files. Existing files re-export from the entity file
so call sites are unchanged.

Refs #563.
